AppApis/app/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0.3
django-admin-volt==0.0.5
django-ckeditor==6.3.2
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==2.0.0
django-oauth-toolkit==1.7.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jwcrypto==1.0
Markdown==3.3.6
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pycparser==2.21
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
zipp==3.8.0
